Refactor URL extraction and processing functionality to improve modularity and maintainability by separating concerns into distinct files. This includes changes to how Discord message content is parsed for stickers, attachments, GIFs, and social media embeds.
Changes:
- Updated the import path for `extractInputs` from `urlExtractor.js` to `urlExtractor/index.js`.
- Increased the limit of fetched messages in `handleNewMessage` from 50 to 100 and modified sorting logic.
- Added system locale handling for timestamp formatting in `handleNewMessage`.
- Enhanced memory block display in `generateReply` to support more entries (from 10 to 15).
- Deleted the old `urlExtractor.ts` file and created multiple new files for improved organisation:
  - `extractDiscord.ts` for handling stickers and attachments.
  - `extractGifs.ts` for fetching GIFs from Tenor and Giphy.
  - `extractInlineImages.ts` for processing inline images.
  - `extractSocialEmbeds.ts` for social media link processing.
  - `index.ts` for aggregating all extraction functionalities.
- Implemented detailed logging across all new modules to track processing steps.
